services/analyzer/requirements.txt
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
mariadb==1.0.11
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
